src/pyproject.toml
[tool.poetry]
name = "spider"
version = "0.1.0"
description = ""
authors = ["keita.oouchi <keita.oouchi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
selenium = "^3.141.0"
google-cloud-storage = "^1.24.1"
Flask = "^1.1.1"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
